This project is designed for button control.
This project is designed to implement button control using the STM32 microcontroller. It aims to create a control system where the digital input pins of the STM32 are used to read a basic button input. Buttons, which are simple input devices commonly used for user interaction, are read by the STM32 microcontroller in this project to control appropriate outputs.

The main components of the project are:

Button Input: Buttons are connected to the digital input pins of the microcontroller and are used to detect whether the user has pressed the button.
GPIO (General Purpose Input/Output): The STM32's GPIO pins read the button inputs and process the signals that trigger specific actions.
Control and Response: When the button is pressed, the system uses logic in the software to control the appropriate output, which could be an LED, motor, alarm, or another device.